+++ install/update_0782_080.php
@@ -951,7 +951,7 @@ function update0782to080() {
             $value = $DB->result($result, 0, 0);
             if (!empty($value)) {
                $query = "UPDATE `glpi_configs`
-                         SET `proxy_passwd` = '".addslashes(Toolbox::encrypt($value, GLPIKEY))."'
+                         SET `proxy_passwd` = '".addslashes(Toolbox::encrypt($value))."'
                          WHERE `id` = '1' ";
                $DB->queryOrDie($query, "0.80 update proxy_passwd in glpi_configs");
             }
@@ -973,7 +973,7 @@ function update0782to080() {
             $value = $DB->result($result, 0, 0);
             if (!empty($value)) {
                $query = "UPDATE `glpi_configs`
-                         SET `smtp_passwd` = '".addslashes(Toolbox::encrypt($value, GLPIKEY))."'
+                         SET `smtp_passwd` = '".addslashes(Toolbox::encrypt($value))."'
                          WHERE `id` = '1' ";
                $DB->queryOrDie($query, "0.80 update smtp_passwd in glpi_configs");
             }
@@ -996,8 +996,7 @@ function update0782to080() {
             while ($data = $DB->fetchAssoc($result)) {
                if (!empty($data['rootdn_password'])) {
                   $query = "UPDATE `glpi_authldaps`
-                            SET `rootdn_passwd` = '".addslashes(Toolbox::encrypt($data['rootdn_password'],
-                                                                                 GLPIKEY))."'
+                            SET `rootdn_passwd` = '".addslashes(Toolbox::encrypt($data['rootdn_password']))."'
                             WHERE `id` = '".$data['id']."' ";
                   $DB->queryOrDie($query, "0.80 update rootdn_passwd in glpi_authldaps");
                }
@@ -1123,8 +1122,7 @@ function update0782to080() {
             while ($data = $DB->fetchAssoc($result)) {
                if (!empty($data['password'])) {
                   $query = "UPDATE `glpi_mailcollectors`
-                            SET `passwd` = '".addslashes(Toolbox::encrypt($data['password'],
-                                                                          GLPIKEY))."'
+                            SET `passwd` = '".addslashes(Toolbox::encrypt($data['password']))."'
                             WHERE `id`= '".$data['id']."' ";
                   $DB->queryOrDie($query, "0.80 update passwd in glpi_mailcollectors");
                }